requirements.txt
grpcio
requests
protobuf
pyyaml
kubernetes
